finish game

+++ CPong/CPong/Ball.h
@@ -4,15 +4,15 @@
 class Ball
 {
 private:
-	int _speed;
+	float _speed;
 	Vector2D _direction;
 	Vector2D _position;
 public:
 	Ball();
 	~Ball();
 
-	void setSpeed(int speed);
-	int getSpeed();
+	void setSpeed(float speed);
+	float getSpeed();
 	void setDirection(Vector2D direction);
 	Vector2D getDirection();
 	void setPosition(Vector2D position);
