+++ CPong/CPong/Vector2D.h
@@ -2,20 +2,20 @@
 class Vector2D
 {
 private:
-	int _x;
-	int _y;
+	float _x;
+	float _y;
 public:
 
 
 	Vector2D();
 	~Vector2D();
-	Vector2D(int x, int y) : _x(x), _y(y){}
+	Vector2D(float x, float y) : _x(x), _y(y){}
 
-	void setX(int x){ _x = x; }
-	int getX(){ return _x; }
+	void setX(float x){ _x = x; }
+	float getX(){ return _x; }
 
-	void setY(int y){ _y = y; }
-	int getY(){ return _y; }
+	void setY(float y){ _y = y; }
+	float getY(){ return _y; }
 
 };
